Set version to be 0.1.0 - it makes more sense to increment te minor revision, rather than the patch level.
This would be 0.1.0 instead of 0.0.3. Allowing different types of comments is a bigger change.

The published extension in the vscode marketplace is 0.0.2, but it seems that the github repository is already ahead of the published version, with the author having added support for descriptions which has also not yet been released.

Signed-off-by: Noscere <noscere1978@gmail.com>