Re-factor the fileheader settings so that the templates are grouped together by the settings UI. It makes more sense for the templates to be the parent field, instead of each language type having a template property (shell.tpl, vb.tpl, etc).
Template descriptions contain markdown (bold) but this is not rendered because the property to be set  needs to be "markdownDescription" and not "description".

Added logic to use the author name in the event that the modified by name is still the default value or has been made blank. The settings UI explains this behaviour and when it will use author name instead.

Added the email property into the configuration settings. This had been missed when functionality for emails was added.