New unit test for fileheader module replaceSubstringInLine().
4 test cases:
Test replacing the author name with a regex.
Test replacing the modified at date & time with a regex.
Test attempting to replace the author name when there is no name in the original text.
Test attempting to reploace the time & date when there is no time or date in the original text.