Initial version of build-licenses-list; added to .vscodeignore so that the script is not included in published package.
Builds a list of licenses available from SPDX database. Parameters allow filtering based on whether a license is deprecated, is FSF free software , and is OSI approved.

Optionally creates a set of contributes enum properties for pasting into the package.json for the extensions settings (allow a user to insert a license into the file copyright header).